backend/requirements.txt
# FastAPI and ASGI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.30.0
alembic>=1.14.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0

# Validation and Settings
pydantic>=2.9.0
pydantic-settings>=2.6.0
email-validator>=2.2.0

# Task Queue
celery[redis]>=5.6.0  # Newer for better 3.14 support
redis>=5.2.0

# ML Libraries
fairlearn>=0.11.0
scipy>=1.16.1
scikit-learn>=1.6.0  # Key upgrade: wheels available for 3.14
pandas>=2.2.3
numpy>=2.0.0  # If issues, change to >=1.26.4,<2.0 later
matplotlib>=3.9.0
seaborn>=0.13.2
joblib>=1.4.0

# Explainability Libraries
shap>=0.46.0  # Better compatibility
lime>=0.2.0.1

# Privacy Libraries
diffprivlib>=0.6.5

# MLflow for experiment tracking
mlflow>=2.16.0  # Fixes old issues

# Deep Learning (uncomment if needed)
# tensorflow>=2.17.0
# torch>=2.5.0
onnxruntime>=1.20.1

# Utilities
python-dotenv>=1.0.1
aiofiles>=24.1.0
setuptools>=75.0.0